a4

+++ src/Kambaz/Courses/Assignments/AssignmentControlButtons.tsx
@@ -1,6 +1,6 @@
 import { IoEllipsisVertical } from "react-icons/io5";
 import GreenCheckmark from "../Modules/GreenCheckmark";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { FaTrash } from "react-icons/fa";
 import { deleteAssignment } from "./reducer";
 
@@ -13,7 +13,12 @@ export default function AssignmentControlButtons({ assignmentId }: {assignmentId
             <IoEllipsisVertical className="fs-4" />
             <FaTrash className="text-danger me-2 mb-1" onClick={(e) => {
                 e.preventDefault();
-                dispatch(deleteAssignment(assignmentId));
+                const check = window.confirm(
+                    "Are you sure you want to delete this assignment?"
+                );
+                if (check) {
+                    dispatch(deleteAssignment(assignmentId));
+                }
             }}/>
         </div>
     );}
